requirements.txt
# --- Core utilities ---
argparse
pathlib
requests==2.32.5

# --- PDF & Image extraction ---
pdfplumber==0.11.7          # text extraction from PDFs
PyMuPDF==1.26.5             # (fitz) extract images & metadata
pypdfium2==4.30.0           # optional PDF rendering backend
pillow==12.0.0              # image processing

# --- OCR layer ---
pytesseract==0.3.13         # Tesseract OCR bridge (requires system tesseract)

# --- Data handling ---
jsonschema==4.25.1
pandas>=2.2.0

# --- Optional NLP / QA generation (if later steps use it) ---
nltk>=3.9
transformers>=4.44.0
torch>=2.4.0
tqdm>=4.66.0

# --- Misc utilities ---
certifi>=2025.10.5
charset-normalizer>=3.4.4
urllib3>=2.5.0
